Add files via upload
I have recently started to learn JAVA since October 2024 and having been 4 months of learning i have been coding up some projects to add to my portfolio. This one specifically is called the F1 race track. The code contains basic attributes such as IF/else statements, for/while loops, Arrays and ArrayLists. 

In a short description complementing the result of the execution shows that each car is given a different amount of increased speed. Once a car reaches the end of the finish line, the number displayed on the top of the car will increase every time it hits the line first. You can then press 's' to restart the race.